Add GetPut and Some Refactor

+++ lib/components/notifications/local_notifications/local_notifications_repository.dart
@@ -100,7 +100,7 @@ class AppLocalNotificationsRepository {
     String? largeIcon,
     bool? locked,
     Map<String, String?>? payload,
-    int? progress,
+    double? progress,
     bool? roundedBigPicture,
     bool? roundedLargeIcon,
     bool? showWhen,
